Fix NPE for dependency methods/groups
When a test method depends on a group/method
which doesn't exist and when a user specifies
[ignoreMissingDependencies=true] instead of running
the test method, TestNG throws a NullPointerException
Fixed this issue.

Also updated the .gitignore file to include 
class path and project eclipse files.

Also removed the class path and project files
so that from next time this issue doesn't happen
to others