Increase randomness of context switches for test
Result of TrueParallelTest is inherently nondeterministic, on systems
with high number of processors, the default number 5 of concurrent running
threads is too low to make the test pass. On my 8 cores system, the test
can rarely pass. This patch changes the number of threads to 4 times of
available processors to make the test more likely to pass. But still
the result of this test is nondeterministic.