Release parameter after use
Parameters saved in InvokedMethod and TestResult make them not able to
be collected after running test even when JVM is running out of memory.
The saved parameters are served as input for report which doesn't really
have to have accurate parameter string. This patch save soft reference to
parameter instead of strong reference, so when JVM is running out of
memory, the original parameter is collected, and the summary of
parameter is used instead in toString().