Fixed dependencies for 4 targets in build.xml
The targets 'tests', 'test-ant', and 'javadocs' all fail without running the
target 'build' first, since they rely on the testng.jar to run. The target
'build' has been added as a dependency to each of these.

The target 'tests' had previously listed 'compile' as a dependency, but this
has been removed since 'build' covers this already.

Another target, 'create-jar', requires the directory target/classes in order
not to fail, and requires the testng classes to make a meaningful jar. It works
properly once the target 'compile' has been run, so 'compile' is now added to
its dependency list.