Standardized section heading HTML format for easier reading/maintenance.
The HTML for the section headings varied between headings. Sometimes
the <h> tags were on surrounding lines, sometimes they wrapped the
content on a single line, and sometimes they were at the end of the
previous line of content. This made for difficult reading; it also made
changes difficult.

Now, all section heading <h> tags wrap the content on a single line,
in the following format:

<hx><a ...>content</a></hx>

so that they are easy to work with.

NOTE: I'm aware that this may not be compliant with some style guides,
and also that everybody has different preferences. I'm not trying to
enforce my particular preference over anyone else's; I picked this
format simply because A) section heading lines are generally short
enough to fit in screen-width even with <h> tags, and B) many of the
section headings were this way to begin with.