AfterClass does not get executed when MethodInterceptor is involved.
The logic to invoke AfterClass was ignoring the output
of MethodInterceptor (especially the ones that alter the
size of the total test methods to be executed). This 
led to AfterClass methods not getting executed at all.
Fixed this problem by ensuring that the classMethodMap
object is updated to refer to the output of a method
interceptor in a TestRunner.