Remove preserveOrderAnna from PreserveOrderTest
Annotation-driven discovery of test methods is backed by
Class#getDeclaredMethods, which returns methods in an undefined order.
This is explained by the Javadoc for that method, which states that "The
elements in the returned array are not sorted and are not in any
particular order." Consequently, this test has undefined behavior and
may pass or fail nondeterministically, which is exactly what I've been
observing with various releases of JDK8.

Fixes #582.