SuiteHTMLReporter: Write files completely
Before this, the SuiteHTMLReporter would write the chronologically list of
methods one method at a time. After every method, the target file was opened and
the new line was written. This behavior was introduced in commit 17c52218805e to
fix an OOM error with many tests.

This commit instead creates a BufferedWriter for the target file and writes the
data directly to the target file, without re-opening it in a loop.

For a random, non-representative test suite with about 2800 tests (the reason
why I am looking into this), this speeds up the SuiteHTMLReporter from about 6.5
seconds to a quarter of a second.

Signed-off-by: Uli Schlachter <psychon@znc.in>