Enable URI format for suite location.
Closes #212

Till now TestNG provided support only for local 
file system paths as suite file locations.

Now enabling the location to be specified as a URI.
TestNG will still support only “file:” URI and 
regular local file system locations, but if one wants
to work with a “http” URI as the suite file location
then only can provide the location as a URL and then
plug-in an implementation of “org.testng.xml.ISuiteParser”
via service loaders and wire in their support for the same.