Exception in DataProvider doesn't fail test run
Closes #217

Following is the gist of changes done:
* Deprecated the nested ExitCodeListener and added
a new one.
* Added null checks to ensure that an ITestResult
object created based on a Description object is not
null. Description Object is typically seen to be null
when JUnit encounters problems in parsing a method 
as a test method.
* The new ExitCodeListener now considers pass/fail/
skip methods through-out the entire suite of suites
to determine the effective return code (or) status.
So altered some tests to now assert on the new status
* Removed an invalid JUnit @Test (annotation on a 
static method is not valid in JUnit world)
* Since new ExitCodeListener is a reporter backed
listener, altered a test that asserts on the reporter
count. (UniqueReporterInjectionTest.java)