DependsOnMethods not respected on the sequential run on 2nd test with same base testClass
Closes #1648

The test DynamicGraphHelperTest had to be fixed
to ensure that it associates instances to which 
Test method belong to, to ensure that the tests 
pass because, they were not packing instances info
and this was causing failures. These tests are not
creating the entire DynamicGraph infrastructure in 
a proper way, but resort to a lot of mocks and stubs.