Fix topologicalSort for more efficiency
Closes #1077 and #1081


This fix reduces the scope of MethodHelper 
toplologicalSort to be more efficient. 
The search scope for "methoddependedupon" was 
changed from all methods to just methods that 
share the same instance. 
If a method has an instance of null then all methods 
are still searched. In my testing this provided a 
massive decrease in load time for large suites 
that group by method.