Add files via upload
Fixed toplologicalSort to be more efficient for depends on method sorting

This is a fix for issues 1077 and 1081.
This fix reduces the scope of MethodHelper toplologicalSort to be more efficient. The search scope for "methoddependedupon" was changed from all methods to just methods that share the same instance. If a method has an instance of null then all methods are still searched. In my testing this provided a massive decrease in load time for large suites that group by method.

Fixes # .

### Did you remember to?

- [ ] Add test case(s)
- [ ] Update `CHANGES.txt`

We encourage pull requests that:

* Add new features to TestNG (or)
* Fix bugs in TestNG

If your pull request involves fixing SonarQube issues then we would suggest that you please discuss this with the 
[TestNG-dev](https://groups.google.com/forum/#!forum/testng-dev) before you spend time working on it.