Re-order to ensure built-in interceptor added first
Closes #1726

Currently TestNG injects the built in method 
interceptor [order preserving interceptor or the 
instance ordering interceptor] as the last. 
So if a user has their own method interceptor 
which attempts at re-ordering the methods, it still
doesn’t have any effect, because we end up ordering
methods finally based on either instances or order
of occurrence of class in the suite xml.

Fixed this by having the built in injectors added 
at the top rather than at the end.

But for custom interceptors that intend to change
method order to work, one would need to ensure that 
preserve-order is DISABLED (this is enabled by default
at suite and test level)