Streamline function name in JS from suite name
Closes #298

The html report in TestNG basically relies on the
suite name when it comes to creating a function
name. But there are rules associated with function
names in JS and it cannot contain special characters.

Fixed this discrepancy by removing off all 
non-alphanumeric characters from the suite name
and replacing them with “_”