Fix #1863 patchset3: remove the blank methods since default methods already implemented from org.testng.ITestListener patchset2: fix the comments from juherr
IMethodInterceptor will be invoked twice when listener implements both
ITestListener and IMethodInterceptor via eclipse execution way.

The root cause is when using ITestRunnerFactory2 to create a new
testRunner, all listeners which implements ITestListener will be added, if this listener is also implements IMethodInterceptor, then it will be added to m_methodInterceptors, however, the implemented IMethodInterceptor will be added to m_methodInterceptors again.

in the second time of adding implemented IMethodInterceptor to
m_methodInterceptors, we add statement to check if this interceptor already exist or not could solve this issue.