Prevent circular dependency
Closes #1874

Currently XmlClass includes both “included” and 
“excluded” methods as part of its equals() and 
hashCode() calculation. 

So in scenarios which involve two XmlClass objects
for the same Class but with different included and 
excluded methods, TestNG seems the two XmlClass
objects as different. 

So when users basically add the same class twice 
but with different methods, TestNG ends up looking
at this as if there’s a cyclic dependency (which is
incorrect).