Test Timeout not respected-parallel="methods" mode
Closes #2009

When TestNG runs parallel tests without an 
executor monitoring the threads, the timeouts 
defined in the test methods aren’t honored and 
TestNG ends up running the tests till it runs to
completion and then uses the timeout value to 
mark the test as failed.

Fixed this by running the test only for as long as
the timeout has been defined after which the current
thread (which is where the test is running as well)
is interrupted.