Streamline instance based parallelism
Combination : parallel=instances, group-by-instances=false

When a user specifies 
parallel=“instances” and 
group-by-instances=“false” we would like to retain
the same behavior as what is seen when user 
specifies parallel=“methods”.

Here the only difference that is there is that
when user specifies parallel=instances, the user 
is only instructing TestNG how to create test class
instances. 

This should not have any effect on the parallelism.