Enable customization of testng-results.xml
Closes #2171

In order for one to be able to generate their 
own customized version of testng-results.xml
without a lot of code duplication, the following
should be done.

1. Extend org.testng.reporters.XMLReporter
2. Override “fileName()” method and provide your
own file name.
3. Override “addCustomTagsFor()” method and plugin
your implementation which will add one or more tags
to the xml file.
4. Add the newly created class as one of the listeners
via “@Listeners” annotation or via “<listeners>” 
tag or via SPI approach.


For a sample of how the customized listener 
can look like, take a look at the sample 
that resides in 
“src/test/java/test/reports/issue2171/MyExampleListener.java”