Streamline parameter initialization
Closes #2238

TestNG reads the JVM arguments to find if a 
parameter had a value, “if and only if” the parameter
was not defined in the testng suite xml file.

It should read the JVM argument all the time and use
the JVM argument’s value as the highest priority
value.

Fixed this discrepancy as part of this PR.