Support DTD redirects when parsing xml files
Currently TestNG DTD is hosted behind a https URL.
TestNG also has setup a direction such that a reference
to the DTD http://testng.org/testng-1.0.dtd automatically
redirects the user to https://testng.org/testng-1.0.dtd.

But the problem here is that apache xerces library
which is what TestNG uses (comes baked into the JDK now)
does not support resolution of entities which 
involves a redirection.

This PR addresses this by adding up a redirection
aware entity resolver so that when the DTD hosting
server does a redirection, TestNG client code can 
follow the redirections and resolve the entity 
properly.