Allow usage of custom DTDs for Suite parsing.
Currently TestNG loads the DTD from the jar only
if its known to it i.e., it should match with one 
of the below URLs:

1. https://beust.com/testng/testng-1.0.dtd
2. https://testng.org/testng-1.0.dtd

If a user provides a custom DTD that is available
in the CLASSPATH, then we dont load it at all.
We always try to go retrieve the one from the internet
Fixed this by looking for the custom DTD file
specified by the user in the class path and if its
not found fall back to the URL to retrieve it.