Streamline DTD Loading
Closes #2403

Following is the logic followed in the PR:

1. If the DTD URL either refers to “testng.org”
(or) “beust.com” then always resort to loading
the embedded DTD.
2. If (1) fails, then just fall back to attempting
to load the DTD from the TestNG site (use “https”
protocol)
3. If the DTD URL is not known to us (i.e.,
is neither “testng.org” nor “beust.com”) then we
check if the user is using “http” protocol.
4. If http protocol is being used and if the JVM
argument [-Dtestng.dtd.http=true] is missing then
throw the error, else load the URL.
5. If its a https protocol then just load the DTD
as is.