Don’t use null objects
Currently when dealing with a BeforeTest config 
method, TestNG quietly slips in a null instance 
into the set. The Set interface states that it 
would throw a NPE if a null value is inserted but 
HashSet (which is what we use) seems to allow null
values into it.

So changing this to start dealing with a special 
object which represents a null.