Favour FutureTask for concurrency support
TestNG currently has a custom implementation of 
ThreadPoolExecutor named GraphThreadPoolExecutor.

This class was created to facilitate concurrency
in a DAG.

Now that we are on JDK11, we can very well move
over to leveraging FutureTask based implementations
and thus decouple ourselves from the Executor 
and just focus on orchestrating the next node retrieval
for execution.

Since this is experimental, we are currently providing
a JVM based switch that can fall back to the old 
Behaviour in case of any issues.

JVM argument to use “-Dtestng.favor.custom.thread-pool.executor=true”