chore: polish build scripts, use Java 17 for the build, target Java 11
Previously, TestNG was build with the "current Java" version which was not very reproducible.
Now the build script would use Java 17 for building the code no matter which Java is selected for
launching Gradle.

Even though the code is always build with Java 17, the tests execute with a configurable Java version.
It enables executing tests with EA Java versions.