Adding issue management workflows
We can now add the below labels to issues 
And the workflow will automatically add the required
text that will be instructions to the issue submitter.

* `needs-sample`
* `question` 
* `help-wanted`

These labels should be added to the repository 
And we can perhaps get rid of the other labels
that intend to convey the same thing.

More details available at: https://github.com/peaceiris/actions-label-commenter