Fix PlainOsgiTest
This test has recently started failing after the PR
https://github.com/testng-team/testng/pull/3089
got merged. 

With the below exception

org.ops4j.pax.exam.TestContainerException: java.lang.IllegalArgumentException: Can not set java.util.List field org.testng.internal.NoOpTestClass.m_beforeTestMethods to [Lorg.testng.ITestNGMethod;

This PR changes the type of a field
Which was previously an array to a list.

The PaxExam listener is making 
Use of reflection to access the data member 
“m_beforeTestMethods” instead of it just using
The “setter” that is part of NoOpTestClass.

Since this library is being used ONLY for test and 
Since the library in question seems to be something
That is not updated for quite sometime, resorting 
to class path overriding wherein we duplicate the 
Same class into TestNG codebase and then fix the 
Problem locally so that the tests will pass.