Allow users to access factory method params info
Closes #3111

TestNG now exposes the IParameterInfo interface 
Via which you can extract the following details
Pertaining to a factory powered test.

* the index - which would match with what was 
Specified in the “indices” attribute of the “Factory” 
Annotation. If nothing was specified, then this
Would be equal to a running count on the total instances.

* current index - which represents a running count
On the total instances.

* The parameters of the factory method

* The instance that was produced.